+++ @xen-orchestra/fs/src/fs.test.js
@@ -33,19 +33,22 @@ const rejectionOf = p =>
     },
     reason => reason
   )
-
+const skipFsNotInAzure = () => {
+  return !!process.env.xo_fs_azure
+}
 // TODO : add tests on encrypted remote
 const handlers = [`file://${tmpdir()}`]
 if (process.env.xo_fs_nfs) handlers.push(process.env.xo_fs_nfs)
 if (process.env.xo_fs_smb) handlers.push(process.env.xo_fs_smb)
+if (process.env.xo_fs_azure) handlers.push(process.env.xo_fs_azure)
 
 handlers.forEach(url => {
   describe(url, () => {
     let handler
 
     const testWithFileDescriptor = (path, flags, fn) => {
-      it('with path', () => fn({ file: path, flags }))
-      it('with file descriptor', async () => {
+      it('with path', { skip: skipFsNotInAzure() }, () => fn({ file: path, flags }))
+      it('with file descriptor', { skip: skipFsNotInAzure() }, async () => {
         const file = await handler.openFile(path, flags)
         try {
           await fn({ file })
@@ -84,11 +87,11 @@ handlers.forEach(url => {
         info = await handler.getInfo()
       })
 
-      it('should return an object with info', async () => {
+      it('should return an object with info', { skip: skipFsNotInAzure() }, async () => {
         assert.equal(typeof info, 'object')
       })
 
-      it('should return correct type of attribute', async () => {
+      it('should return correct type of attribute', { skip: skipFsNotInAzure() }, async () => {
         if (info.size !== undefined) {
           assert.equal(typeof info.size, 'number')
         }
@@ -128,7 +131,7 @@ handlers.forEach(url => {
     })
 
     describe('#mkdir()', () => {
-      it('creates a directory', async () => {
+      it('creates a directory', { skip: skipFsNotInAzure() }, async () => {
         await handler.mkdir('dir')
         assert.deepEqual(await handler.list('.'), ['dir'])
       })
@@ -146,24 +149,24 @@ handlers.forEach(url => {
     })
 
     describe('#mktree()', () => {
-      it('creates a tree of directories', async () => {
+      it('creates a tree of directories', { skip: skipFsNotInAzure() }, async () => {
         await handler.mktree('dir/dir')
         assert.deepEqual(await handler.list('.'), ['dir'])
         assert.deepEqual(await handler.list('dir'), ['dir'])
       })
 
-      it('does not throw on existing directory', async () => {
+      it('does not throw on existing directory', { skip: skipFsNotInAzure() }, async () => {
         await handler.mktree('dir/dir')
         await handler.mktree('dir/dir')
       })
 
-      it('throws ENOTDIR on existing file', async () => {
+      it('throws ENOTDIR on existing file', { skip: skipFsNotInAzure() }, async () => {
         await handler.outputFile('dir/file', '')
         const error = await rejectionOf(handler.mktree('dir/file'))
         assert.equal(error.code, 'ENOTDIR')
       })
 
-      it('throws ENOTDIR on existing file in path', async () => {
+      it('throws ENOTDIR on existing file in path', { skip: skipFsNotInAzure() }, async () => {
         await handler.outputFile('file', '')
         const error = await rejectionOf(handler.mktree('file/dir'))
         assert.equal(error.code, 'ENOTDIR')
@@ -176,8 +179,7 @@ handlers.forEach(url => {
         assert.deepEqual(await handler.readFile('file'), TEST_DATA)
       })
 
-      it('throws on existing files', async () => {
-        await handler.unlink('file')
+      it('throws on existing files', { skip: skipFsNotInAzure() }, async () => {
         await handler.outputFile('file', '')
         const error = await rejectionOf(handler.outputFile('file', ''))
         assert.equal(error.code, 'EEXIST')
@@ -242,21 +244,20 @@ handlers.forEach(url => {
     })
 
     describe('#rmdir()', () => {
-      it('should remove an empty directory', async () => {
+      it('should remove an empty directory', { skip: skipFsNotInAzure() }, async () => {
         await handler.mkdir('dir')
         await handler.rmdir('dir')
         assert.deepEqual(await handler.list('.'), [])
       })
 
-      it(`should throw on non-empty directory`, async () => {
+      it(`should throw on non-empty directory`, { skip: skipFsNotInAzure() }, async () => {
         await handler.outputFile('dir/file', '')
 
         const error = await rejectionOf(handler.rmdir('.'))
         assert.equal(error.code, 'ENOTEMPTY')
-        await handler.unlink('dir/file')
       })
 
-      it('does not throw on missing directory', async () => {
+      it('does not throw on missing directory', { skip: skipFsNotInAzure() }, async () => {
         await handler.rmdir('dir')
       })
     })
@@ -265,13 +266,12 @@ handlers.forEach(url => {
       it(`should remove a directory resursively`, async () => {
         await handler.outputFile('dir/file', '')
         await handler.rmtree('dir')
-
-        assert.deepEqual(await handler.list('.'), [])
+        assert.deepEqual(await handler.list('.', { ignoreMissing: true }), [])
       })
     })
 
     describe('#test()', () => {
-      it('tests the remote appears to be working', async () => {
+      it('tests the remote appears to be working', { skip: skipFsNotInAzure() }, async () => {
         const answer = await handler.test()
 
         assert.equal(answer.success, true)
@@ -285,7 +285,7 @@ handlers.forEach(url => {
         await handler.outputFile('file', TEST_DATA)
         await handler.unlink('file')
 
-        assert.deepEqual(await handler.list('.'), [])
+        assert.deepEqual(await handler.list('.', { ignoreMissing: true }), [])
       })
 
       it('does not throw on missing file', async () => {
@@ -348,7 +348,7 @@ handlers.forEach(url => {
           })(),
         },
         ({ length, expected }, title) => {
-          it(title, async () => {
+          it(title, { skip: skipFsNotInAzure() }, async () => {
             await handler.outputFile('file', TEST_DATA)
             await handler.truncate('file', length)
             assert.deepEqual(await handler.readFile('file'), expected)