+++ @xen-orchestra/lite/src/components/host/dashboard/HostDashboardLoadAverageChart.vue
@@ -55,13 +55,14 @@ const loadAverage = computed<LinearChartData>(() => {
 })
 
 const maxValue = computed(() => {
-  const values = loadAverage.value[0]?.data.map(item => item.value) ?? []
+  const values = loadAverage.value[0]?.data.map(item => item.value ?? 0) ?? []
 
   if (values.length === 0) {
     return 10
   }
 
   const maxLoad = Math.max(...values)
+
   return Math.ceil(maxLoad / 5) * 5
 })
 </script>