+++ @xen-orchestra/web-core/lib/components/backup-state/VtsBackupState.vue
@@ -1,29 +1,32 @@
 <template>
-  <VtsIcon :accent="color" :icon class="vts-backup-state" />
+  <UiInfo :accent="state.accent" class="vts-backup-state">
+    {{ state.text }}
+  </UiInfo>
 </template>
 
 <script lang="ts" setup>
-import VtsIcon from '@core/components/icon/VtsIcon.vue'
-import type { IconDefinition } from '@fortawesome/fontawesome-common-types'
-import { faCheckCircle, faCircleMinus, faCircleXmark } from '@fortawesome/free-solid-svg-icons'
-import { computed } from 'vue'
+import UiInfo, { type InfoAccent } from '@core/components/ui/info/UiInfo.vue'
+import { useMapper } from '@core/packages/mapper'
+import { useI18n } from 'vue-i18n'
 
-export type BackupState = 'success' | 'failure' | 'partial'
-
-type Props = {
+const { state: _state } = defineProps<{
   state: BackupState
-}
+}>()
 
-const props = defineProps<Props>()
+const { t } = useI18n()
 
-const states: Record<Props['state'], { icon: IconDefinition; color: 'success' | 'warning' | 'danger' }> = {
-  success: { icon: faCheckCircle, color: 'success' },
-  partial: { icon: faCircleMinus, color: 'warning' },
-  failure: { icon: faCircleXmark, color: 'danger' },
-}
+type BackupState = 'success' | 'failure' | 'skipped' | 'interrupted'
 
-const icon = computed(() => states[props.state].icon)
-const color = computed(() => states[props.state].color)
+const state = useMapper<BackupState, { text: string; accent: InfoAccent }>(
+  () => _state,
+  {
+    success: { text: t('success'), accent: 'success' },
+    failure: { text: t('failure'), accent: 'danger' },
+    skipped: { text: t('skipped'), accent: 'warning' },
+    interrupted: { text: t('interrupted'), accent: 'danger' },
+  },
+  'failure'
+)
 </script>
 
 <style lang="postcss" scoped>