+++ @xen-orchestra/web-core/lib/components/copy-button/VtsCopyButton.vue
@@ -1,13 +1,13 @@
 <template>
-  <UiButtonIcon v-tooltip="copied && t('core.copied')" :icon size="medium" accent="brand" @click="copyToClipboard()" />
+  <UiButtonIcon v-tooltip="copied && t('core.copied')" :icon size="medium" accent="brand" @click="copy()" />
 </template>
 
 <script setup lang="ts">
 import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import { vTooltip } from '@core/directives/tooltip.directive'
 import { faCheckCircle, faCopy } from '@fortawesome/free-solid-svg-icons'
-import { useClipboard, useTimeoutFn } from '@vueuse/core'
-import { ref } from 'vue'
+import { useClipboard } from '@vueuse/core'
+import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 const { value } = defineProps<{
@@ -16,17 +16,7 @@ const { value } = defineProps<{
 
 const { t } = useI18n()
 
-const { copy, copied } = useClipboard()
+const { copy, copied } = useClipboard({ source: () => value })
 
-const icon = ref(faCopy)
-
-const { start: changeIcon } = useTimeoutFn(() => {
-  icon.value = faCopy
-}, 1_500) // 1.5s is time to Tooltip is visible
-
-function copyToClipboard() {
-  copy(value)
-  icon.value = faCheckCircle
-  changeIcon()
-}
+const icon = computed(() => (copied.value ? faCheckCircle : faCopy))
 </script>