+++ packages/xo-server/src/xo-mixins/jobs/index.mjs
@@ -176,6 +176,7 @@ export default class Jobs {
               mode: job.mode,
               reportWhen: job.settings['']?.reportWhen ?? 'failure',
               backupReportTpl: job.settings['']?.backupReportTpl,
+              hideSuccessfulItems: job.settings['']?.hideSuccessfulItems,
             }
           : undefined,
       event: 'job.start',
@@ -276,7 +277,7 @@ export default class Jobs {
       runs[runJobId] = { cancel }
       $defer(() => delete runs[runJobId])
 
-      const status = await executor({
+      await executor({
         app,
         cancelToken: token,
         connection,
@@ -296,7 +297,7 @@ export default class Jobs {
         true
       )
 
-      app.emit('job:terminated', runJobId, { status, type })
+      app.emit('job:terminated', runJobId, { type })
     } catch (error) {
       await logger.error(
         `The execution of ${id} has failed.`,