+++ packages/xo-server/src/xo-mixins/xen-servers.mjs
@@ -130,7 +130,7 @@ export default class XenServers {
   }
 
   async unregisterXenServer(id) {
-    this.disconnectXenServer(id)::ignoreErrors()
+    await this.disconnectXenServer(id)
 
     if (!(await this._servers.remove(id))) {
       throw noSuchObject(id, 'xenServer')
@@ -601,7 +601,14 @@ export default class XenServers {
     if (server.status === 'connected') {
       const xapi = xapis[server.id]
       server.poolId = xapi.pool.uuid
-      server.master = xapi.getObjectByRef(xapi.pool.master).uuid
+      try {
+        server.master = xapi.getObjectByRef(xapi.pool.master).uuid
+      } catch (error) {
+        // Hosts may not be loaded
+        if (!noSuchObject.is(error)) {
+          throw error
+        }
+      }
     }
     if (server.label === undefined) {
       server.label = server.poolNameLabel